privatize other code
