+++ zephyr_ml/__init__.py
@@ -9,8 +9,6 @@ __version__ = '0.0.5.dev10'
 import os
 
 from zephyr_ml.core import Zephyr
-from zephyr_ml.entityset import VALIDATE_DATA_FUNCTIONS, _create_entityset
-from zephyr_ml.labeling import DataLabeler
 
 MLBLOCKS_PRIMITIVES = os.path.join(
     os.path.dirname(__file__), "primitives", "jsons")
